课程进度：技能系统(10)代码，至此技能流程基本补全。下一步就创建具体的技能行为实体。

+++ Unity/Assets/Scripts/Codes/Model/Server/MMO/Battle/Buff/Buff.cs
@@ -37,7 +37,7 @@ namespace ET.Server
 
         public int TickTime;
 
-        public int TickBeginTime;
+        public long TickBeginTime;
 
         [BsonIgnore]
         public long TickTimer;
